\name{motorcycledata}
\docType{data}
\alias{motorcycledata}
\title{
Motorcycle data.
}
\description{This table gives the results of 133 simulations showing the effects of motorcycle crashes on victims heads: time after a simulated impact with motorcycles and head acceleration of a PTMO (post mortem human test object) were recorded.}
\references{Hardle, W. (1990) \emph{Applied Nonparametric Regression}. Cambridge University Press.}
\usage{data(motorcycledata)}
\format{A 133 by 2 data frame.}
\keyword{datasets}


