\name{postmean.cauchy}
\alias{postmean.cauchy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{postmean.cauchy}
\description{
Posterior mean calculations for Bayesian thresholding.
}
\details{
This function replaces one in the EbayesThresh package, which perform Bayesian thresholding.  For more information, see help by Silverman (see references below).
}
\references{
Johnstone, I.M. and Silverman, B.W. (2002) EbayesThresh:  R and S-PLUS software for Empirical Bayes thresholding (Submitted for publication).


Johnstone, I.M. and Silverman, B.W. (2004) Needles and hay in haystacks: Empirical Bayes estimates of possibly sparse sequences.  \emph{Ann. Statist.}, \bold{32}, 1594--1649.
    
}
%\author{   who you are   }
%\note{The new function \code{postmean.cauchy} approximates \code{postmean} at small \env{x} values, since the original function \code{postmean.cauchy} supplied with the EbayesThresh pacakge does not cope well with this. 
%}

\seealso{\code{\link{denoise}},\code{\link{denoisehetero}}, \code{\link{denoiseheteromp}}, \code{\link{denoiseheteroprop}}     }

\keyword{methods}% at least one, from doc/KEYWORDS

